# public/openapi/moim-api.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "MOIM API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://{env}.moim.id/api/v1",
            "variables": {
                "env": {
                    "enum": [
                        "dev",
                        "www"
                    ],
                    "default": "dev"
                }
            }
        }
    ],
...
    "components": {
        "schemas": {
            "ApiResponse": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer"
                    },
                    "success": {
                        "type": "boolean"
                    },
                    "messages": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "paths": {
...
        "/moim": {
            "post": {
                "tags": [
                    "MOIM"
                ],
                "summary": "Untuk membuat MOIM",
                "description": "",
                "parameters": [
                    {
                        "name": "app-token",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {}
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {}
                    }
                ],
                "requestBody": {
                    "description": "data MOIM",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "about": {
                                        "type": "string"
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "category_id": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    },
                                    "photo": {
                                        "type": "object",
                                        "properties": {
                                            "photo_path": {
                                                "type": "string"
                                            },
                                            "avatar_name": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "preferences": {
                                        "type": "object",
                                        "properties": {
                                            "private_moim": {
                                                "type": "boolean"
                                            },
                                            "custom_profil": {
                                                "type": "boolean"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "object",
                                                    "properties": {
                                                        "name": {
                                                            "type": "string"
                                                        },
                                                        "about": {
                                                            "type": "string"
                                                        },
                                                        "category_id": {
                                                            "type": "string"
                                                        },
                                                        "photo": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "photo_path": {
                                                                        "type": "string"
                                                                    },
                                                                    "avatar": {
                                                                        "type": "string"
                                                                    }
                                                                }
                                                            }
                                                        },
                                                        "preferences": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "private_moim": {
                                                                        "type": "boolean"
                                                                    },
                                                                    "custom_profile": {
                                                                        "type": "boolean"
                                                                    }
                                                                }
                                                            }
                                                        },
                                                        "total_post": {
                                                            "type": "integer"
                                                        },
                                                        "totl_post_weekly": {
                                                            "type": "integer"
                                                        },
                                                        "total_members": {
                                                            "type": "integer"
                                                        },
                                                        "members": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "user_id": {
                                                                        "type": "string"
                                                                    },
                                                                    "user_role": {
                                                                        "type": "string"
                                                                    },
                                                                    "member_status": {
                                                                        "type": "string",
                                                                        "format": "requested,joined,rejcted,blocked"
                                                                    },
                                                                    "joined_at": {
                                                                        "type": "string",
                                                                        "format": "datetime"
                                                                    },
                                                                    "deleted_at": {
                                                                        "type": "string",
                                                                        "format": "datetime"
                                                                    }
                                                                }
                                                            }
                                                        },
                                                        "posts": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "post_id": {
                                                                        "type": "string"
                                                                    },
                                                                    "created_at": {
                                                                        "type": "string",
                                                                        "format": "datetime"
                                                                    },
                                                                    "created_by": {
                                                                        "type": "string"
                                                                    },
                                                                    "deleted_at": {
                                                                        "type": "string",
                                                                        "format": "datetime"
                                                                    },
                                                                    "deleted_by": {
                                                                        "type": "string"
                                                                    }
                                                                }
                                                            }
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "datetime"
                                                        },
                                                        "created_by": {
                                                            "type": "string"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "datetime"
                                                        },
                                                        "updated_by": {
                                                            "type": "string"
                                                        },
                                                        "deleted_at": {
                                                            "type": "string",
                                                            "format": "datetime"
                                                        },
                                                        "deleted_by": {
                                                            "type": "string"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/moim": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Untuk menampilkan data grup yang diikuti",
                "description": "",
                "parameters": [
                    {
                        "name": "app-token",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {}
                    },
                    {
                        "name": "app-token",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "id": {
                                                                "type": "string"
                                                            },
                                                            "moim_name": {
                                                                "type": "string"
                                                            },
                                                            "total_post": {
                                                                "type": "integer"
                                                            },
                                                            "profil_moim": {
                                                                "type": "string"
                                                            },
                                                            "last_post": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "description": {
                                                                        "type": "string"
                                                                    },
                                                                    "author_id": {
                                                                        "type": "string"
                                                                    },
                                                                    "author_name": {
                                                                        "type": "string"
                                                                    },
                                                                    "created_at": {
                                                                        "type": "string",
                                                                        "format": "datetime"
                                                                    }
                                                                }
                                                            }
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post":{
                "tags": [
                    "User"
                ],
                "summary": "Untuk mengundang user lain ke dalam MOIM",
                "description": "",
                "parameters": [
                    {
                        "name": "app-token",
                        "in": "header",
                        "description": "",
                        "schema":{}
                    }
                ],
                "responses": {
                    "200":{
                        "description": "",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data":{
                                                    "type": "array",
                                                    "items":{
                                                        "type": "object",
                                                        "properties": {
                                                            "id_moim":{
                                                                "type": "string"
                                                            },
                                                            "moim_name":{
                                                                "type": "string"
                                                            },
                                                            "total_unread_post":{
                                                                "type": "integer"
                                                            },
                                                            "profil_moim":{
                                                                "type": "string",
                                                                "format": "avatar"
                                                            },
                                                            "last_post":{
                                                                "type": "object"
                                                            }
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}